requirements.txt
discord.py
requests
discord-pretty-help
Pillow